Fixing issue with child process termination.
Because I wasn't properly leaving the loop during the fork, child
processes would attempt to run code that should only be executed in the
master - causing none of the children to actually exit correctly.

This means that if a child died due to an error there could be issues
with open file descriptors.